Add some decorator classes to make overrides
