+++ decidim-process-extended/app/decorators/decidim/participatory_processes/participatory_processes_controller_decorator.rb
@@ -1,0 +1,27 @@
+# frozen_string_literal: true
+
+Decidim::ParticipatoryProcesses::ParticipatoryProcessesController.class_eval do
+  def published_processes
+    byebug
+    @published_processes ||= OrganizationPublishedParticipatoryProcesses.new(current_organization, current_user).query.where(decidim_participatory_process_group_id: Rails.application.config.process)
+  end
+
+  # This is customized because GENCAT don't Processes Groups on Index Page
+  def collection
+    byebug
+    @collection ||= participatory_processes
+  end
+  # This is customized because GENCAT don't Processes Groups on Index Page
+
+  def filtered_processes
+    byebug
+    search.results.where(decidim_participatory_process_group_id: Rails.application.config.process)
+  end
+
+  # This is customized because GENCAT
+  def participatory_processes
+    byebug
+    @participatory_processes ||= filtered_processes
+  end
+  # This is customized because GENCAT
+end
